docs(embedding): clarify stable API boundary and update embedding docs
- Mark `configure_container_git_identity` as internal shim, not part of stable API
- Revise and improve users and developers guide for embedding with stable APIs only
- Add compile tests verifying internal feature gating of config re-exports
- Remove most exec helper tests from helpers.rs to helpers_tests.rs for cleaner structure
- Add detailed class diagram and explanations of exec API types for better documentation clarity
